// API.ts
export type AggregatedStats = {
  __typename: "AggregatedStats",
  createdAt: string,
  id: string,
  losses: number,
  metadata?: string | null,
  owner?: string | null,
  statsKey: string,
  statsType: string,
  totalGames: number,
  updatedAt?: string | null,
  winRate: number,
  wins: number,
};

export type Deck = {
  __typename: "Deck",
  className: string,
  createdAt?: string | null,
  deckName: string,
  id: string,
  owner?: string | null,
  updatedAt: string,
  userId: string,
};

export type MatchRecord = {
  __typename: "MatchRecord",
  createdAt: string,
  id: string,
  isFirstPlayer: boolean,
  isWin: boolean,
  myDeckId: string,
  opponentClass: string,
  opponentDeckType: string,
  owner?: string | null,
  recordedAt: string,
  updatedAt: string,
  userId: string,
};

export type User = {
  __typename: "User",
  createdAt?: string | null,
  email: string,
  id: string,
  owner?: string | null,
  updatedAt: string,
};
// ...
export type ModelStringInput = {
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  beginsWith?: string | null,
  between?: Array< string | null > | null,
  contains?: string | null,
  eq?: string | null,
  ge?: string | null,
  gt?: string | null,
  le?: string | null,
  lt?: string | null,
  ne?: string | null,
  notContains?: string | null,
  size?: ModelSizeInput | null,
};
// ...
export type ModelSubscriptionAggregatedStatsFilterInput = {
  and?: Array< ModelSubscriptionAggregatedStatsFilterInput | null > | null,
  createdAt?: ModelSubscriptionStringInput | null,
  id?: ModelSubscriptionIDInput | null,
  losses?: ModelSubscriptionIntInput | null,
  metadata?: ModelSubscriptionStringInput | null,
  or?: Array< ModelSubscriptionAggregatedStatsFilterInput | null > | null,
  owner?: ModelStringInput | null,
  statsKey?: ModelSubscriptionStringInput | null,
  statsType?: ModelSubscriptionStringInput | null,
  totalGames?: ModelSubscriptionIntInput | null,
  updatedAt?: ModelSubscriptionStringInput | null,
  winRate?: ModelSubscriptionFloatInput | null,
  wins?: ModelSubscriptionIntInput | null,
};

export type ModelSubscriptionStringInput = {
  beginsWith?: string | null,
  between?: Array< string | null > | null,
  contains?: string | null,
  eq?: string | null,
  ge?: string | null,
  gt?: string | null,
  in?: Array< string | null > | null,
  le?: string | null,
  lt?: string | null,
  ne?: string | null,
  notContains?: string | null,
  notIn?: Array< string | null > | null,
};

export type ModelSubscriptionIDInput = {
  beginsWith?: string | null,
  between?: Array< string | null > | null,
  contains?: string | null,
  eq?: string | null,
  ge?: string | null,
  gt?: string | null,
  in?: Array< string | null > | null,
  le?: string | null,
  lt?: string | null,
  ne?: string | null,
  notContains?: string | null,
  notIn?: Array< string | null > | null,
};

export type ModelSubscriptionIntInput = {
  between?: Array< number | null > | null,
  eq?: number | null,
  ge?: number | null,
  gt?: number | null,
  in?: Array< number | null > | null,
  le?: number | null,
  lt?: number | null,
  ne?: number | null,
  notIn?: Array< number | null > | null,
};

export type ModelSubscriptionFloatInput = {
  between?: Array< number | null > | null,
  eq?: number | null,
  ge?: number | null,
  gt?: number | null,
  in?: Array< number | null > | null,
  le?: number | null,
  lt?: number | null,
  ne?: number | null,
  notIn?: Array< number | null > | null,
};

export type ModelSubscriptionDeckFilterInput = {
  and?: Array< ModelSubscriptionDeckFilterInput | null > | null,
  className?: ModelSubscriptionStringInput | null,
  createdAt?: ModelSubscriptionStringInput | null,
  deckName?: ModelSubscriptionStringInput | null,
  id?: ModelSubscriptionIDInput | null,
  or?: Array< ModelSubscriptionDeckFilterInput | null > | null,
  owner?: ModelStringInput | null,
  updatedAt?: ModelSubscriptionStringInput | null,
  userId?: ModelSubscriptionIDInput | null,
};

export type ModelSubscriptionMatchRecordFilterInput = {
  and?: Array< ModelSubscriptionMatchRecordFilterInput | null > | null,
  createdAt?: ModelSubscriptionStringInput | null,
  id?: ModelSubscriptionIDInput | null,
  isFirstPlayer?: ModelSubscriptionBooleanInput | null,
  isWin?: ModelSubscriptionBooleanInput | null,
  myDeckId?: ModelSubscriptionIDInput | null,
  opponentClass?: ModelSubscriptionStringInput | null,
  opponentDeckType?: ModelSubscriptionStringInput | null,
  or?: Array< ModelSubscriptionMatchRecordFilterInput | null > | null,
  owner?: ModelStringInput | null,
  recordedAt?: ModelSubscriptionStringInput | null,
  updatedAt?: ModelSubscriptionStringInput | null,
  userId?: ModelSubscriptionIDInput | null,
};

export type ModelSubscriptionBooleanInput = {
  eq?: boolean | null,
  ne?: boolean | null,
};

export type ModelSubscriptionUserFilterInput = {
  and?: Array< ModelSubscriptionUserFilterInput | null > | null,
  createdAt?: ModelSubscriptionStringInput | null,
  email?: ModelSubscriptionStringInput | null,
  id?: ModelSubscriptionIDInput | null,
  or?: Array< ModelSubscriptionUserFilterInput | null > | null,
  owner?: ModelStringInput | null,
  updatedAt?: ModelSubscriptionStringInput | null,
};
// ...
export type OnCreateAggregatedStatsSubscriptionVariables = {
  filter?: ModelSubscriptionAggregatedStatsFilterInput | null,
  owner?: string | null,
};

export type OnCreateAggregatedStatsSubscription = {
  onCreateAggregatedStats?:  {
    __typename: "AggregatedStats",
    createdAt: string,
    id: string,
    losses: number,
    metadata?: string | null,
    owner?: string | null,
    statsKey: string,
    statsType: string,
    totalGames: number,
    updatedAt?: string | null,
    winRate: number,
    wins: number,
  } | null,
};

export type OnCreateDeckSubscriptionVariables = {
  filter?: ModelSubscriptionDeckFilterInput | null,
  owner?: string | null,
};

export type OnCreateDeckSubscription = {
  onCreateDeck?:  {
    __typename: "Deck",
    className: string,
    createdAt?: string | null,
    deckName: string,
    id: string,
    owner?: string | null,
    updatedAt: string,
    userId: string,
  } | null,
};

export type OnCreateMatchRecordSubscriptionVariables = {
  filter?: ModelSubscriptionMatchRecordFilterInput | null,
  owner?: string | null,
};

export type OnCreateMatchRecordSubscription = {
  onCreateMatchRecord?:  {
    __typename: "MatchRecord",
    createdAt: string,
    id: string,
    isFirstPlayer: boolean,
    isWin: boolean,
    myDeckId: string,
    opponentClass: string,
    opponentDeckType: string,
    owner?: string | null,
    recordedAt: string,
    updatedAt: string,
    userId: string,
  } | null,
};

export type OnCreateUserSubscriptionVariables = {
  filter?: ModelSubscriptionUserFilterInput | null,
  owner?: string | null,
};

export type OnCreateUserSubscription = {
  onCreateUser?:  {
    __typename: "User",
    createdAt?: string | null,
    email: string,
    id: string,
    owner?: string | null,
    updatedAt: string,
  } | null,
};

export type OnDeleteAggregatedStatsSubscriptionVariables = {
  filter?: ModelSubscriptionAggregatedStatsFilterInput | null,
  owner?: string | null,
};

export type OnDeleteAggregatedStatsSubscription = {
  onDeleteAggregatedStats?:  {
    __typename: "AggregatedStats",
    createdAt: string,
    id: string,
    losses: number,
    metadata?: string | null,
    owner?: string | null,
    statsKey: string,
    statsType: string,
    totalGames: number,
    updatedAt?: string | null,
    winRate: number,
    wins: number,
  } | null,
};

export type OnDeleteDeckSubscriptionVariables = {
  filter?: ModelSubscriptionDeckFilterInput | null,
  owner?: string | null,
};

export type OnDeleteDeckSubscription = {
  onDeleteDeck?:  {
    __typename: "Deck",
    className: string,
    createdAt?: string | null,
    deckName: string,
    id: string,
    owner?: string | null,
    updatedAt: string,
    userId: string,
  } | null,
};

export type OnDeleteMatchRecordSubscriptionVariables = {
  filter?: ModelSubscriptionMatchRecordFilterInput | null,
  owner?: string | null,
};

export type OnDeleteMatchRecordSubscription = {
  onDeleteMatchRecord?:  {
    __typename: "MatchRecord",
    createdAt: string,
    id: string,
    isFirstPlayer: boolean,
    isWin: boolean,
    myDeckId: string,
    opponentClass: string,
    opponentDeckType: string,
    owner?: string | null,
    recordedAt: string,
    updatedAt: string,
    userId: string,
  } | null,
};

export type OnDeleteUserSubscriptionVariables = {
  filter?: ModelSubscriptionUserFilterInput | null,
  owner?: string | null,
};

export type OnDeleteUserSubscription = {
  onDeleteUser?:  {
    __typename: "User",
    createdAt?: string | null,
    email: string,
    id: string,
    owner?: string | null,
    updatedAt: string,
  } | null,
};

export type OnUpdateAggregatedStatsSubscriptionVariables = {
  filter?: ModelSubscriptionAggregatedStatsFilterInput | null,
  owner?: string | null,
};

export type OnUpdateAggregatedStatsSubscription = {
  onUpdateAggregatedStats?:  {
    __typename: "AggregatedStats",
    createdAt: string,
    id: string,
    losses: number,
    metadata?: string | null,
    owner?: string | null,
    statsKey: string,
    statsType: string,
    totalGames: number,
    updatedAt?: string | null,
    winRate: number,
    wins: number,
  } | null,
};

export type OnUpdateDeckSubscriptionVariables = {
  filter?: ModelSubscriptionDeckFilterInput | null,
  owner?: string | null,
};

export type OnUpdateDeckSubscription = {
  onUpdateDeck?:  {
    __typename: "Deck",
    className: string,
    createdAt?: string | null,
    deckName: string,
    id: string,
    owner?: string | null,
    updatedAt: string,
    userId: string,
  } | null,
};

export type OnUpdateMatchRecordSubscriptionVariables = {
  filter?: ModelSubscriptionMatchRecordFilterInput | null,
  owner?: string | null,
};

export type OnUpdateMatchRecordSubscription = {
  onUpdateMatchRecord?:  {
    __typename: "MatchRecord",
    createdAt: string,
    id: string,
    isFirstPlayer: boolean,
    isWin: boolean,
    myDeckId: string,
    opponentClass: string,
    opponentDeckType: string,
    owner?: string | null,
    recordedAt: string,
    updatedAt: string,
    userId: string,
  } | null,
};

export type OnUpdateUserSubscriptionVariables = {
  filter?: ModelSubscriptionUserFilterInput | null,
  owner?: string | null,
};

export type OnUpdateUserSubscription = {
  onUpdateUser?:  {
    __typename: "User",
    createdAt?: string | null,
    email: string,
    id: string,
    owner?: string | null,
    updatedAt: string,
  } | null,
};